Moved locking behaviour into common superclass. Moved context-managed behaviour into common superclass. Implemented bitbangio.SPI busio.SPI, busio.I2C, busio.UART
